FastSpeech2 first impl

+++ dataloader/dataset.py
@@ -30,9 +30,9 @@ def get_data_to_buffer():
         character = text[i][0:len(text[i])-1]
         character = np.array(text_to_sequence(character, ['english_cleaners']))
         pitch = np.load(os.path.join("./data/pitch", "ljspeech-pitch-%05d.npy" % (i+1)))
-        pitch = torch.from_numpy(pitch)
+        pitch = torch.from_numpy(pitch).to(torch.float32)
         energy = np.load(os.path.join("./data/energy", "ljspeech-energy-%05d.npy" % (i+1)))
-        energy = torch.from_numpy(energy)
+        energy = torch.from_numpy(energy).to(torch.float32)
 
         character = torch.from_numpy(character)
         duration = torch.from_numpy(duration)
